Variable fix

+++ Blagajna/Form1.cs
@@ -165,7 +165,7 @@ namespace Blagajna
                 }
             }
 
-            if (Int32.Parse(kolicinaTB.Text) < 0)
+            if (Double.Parse(kolicinaTB.Text) < 0)
             {
                 potrebnoPoljeLabel.Text = "*Kriva vrijednost";
                 return;
